update Pently to 0.05wip12
Pently 5 is here used in a configuration comparable to Pently 3,
which adds legato, grace note, mid-pattern instrument change, and
other effects not used by this game's Pently 2 soundtrack.

PRG size increases by about 200 bytes to 10.6 KiB.  But it's worth
it to reduce the number of projects in the wild relying on obsolete
Pently versions, as well as for inclusion in multicarts.