title screen to stripe image
I don't plan for the Apple PackBits decoder to be part of the
library long term.  In particular, it won't be part of a
multicart of my collected works, whereas PB53 (unary RLE) and
Popslide (an NES Stripe Image interpreter) will be.

Switching to Popslide also opens up more shuffle possibilities
for the title screen's data.